Fix log4j shutdown hook FATAL log. Added manifest attributes to main jar. Updated initialization of manifest attribute Map and its dependencies. Added checking on closing resources.

+++ src/main/java/com/aaron/desktop/model/others/ApplicationLock.java
@@ -59,9 +59,20 @@ public final class ApplicationLock
     {
         try
         {
-            this.lock.release();
-            this.channel.close();
-            this.lockFile.delete();
+            if(this.lock != null && this.lock.isValid())
+            {
+                this.lock.release();
+            }
+
+            if(this.channel != null)
+            {
+                this.channel.close();
+            }
+
+            if(this.lockFile != null && this.lockFile.exists())
+            {
+                this.lockFile.delete();
+            }
         }
         catch (IOException ex)
         {
